fix(1448-s3-monitoring-does-not-work): fix image size, add file upload during init

+++ src/db/include/awsmock/entity/apps/Application.h
@@ -46,12 +46,12 @@ namespace AwsMock::Database::Entity::Apps {
         /**
          * Private port. This is the internal port used by the application inside the docker image.
          */
-        long privatePort{};
+        int privatePort{};
 
         /**
          * Public port. This is the host port used by the application inside the docker image.
          */
-        long publicPort{};
+        int publicPort{};
 
         /**
          * Name of the container
